make dns.py compatible with ansible 1.x and 2.x
Trusty ships with ansible 1, xenial with ansible 2. The lookup
plugin format changed with ansible 2, so we need to modify dns.py
to be compatible with both versions.

Searching led me to a few examples of hybrid plugins:

http://47.95.33.19:8080/ansible_alicloud/latest/porting_guide_2.0.html#id2
https://github.com/jhaals/ansible-vault/blob/master/vault.py#L23

I'm not ansible guru, but this seemed easy enough to implement for
dns.py. I've successfully tested this change with both trusty and
xenial deployments of the elasticsearch charm.