Fix unit status after installation
The unit status is always being shown as blocked after the installation
of the charm. This happens because the hook was testing for the shards
of the node which right after the installation will be 0.

Now we just check if the local endpoint is available, meaning that the
service is ready to be accessed.

Closes-bug: #1827895